add svg_stack tool from https://github.com/astraw/svg_stack
This tool allows to concatenate svg files. I have tried to include
the actual repository as a submodule, but it is a huge pain in the
ass as it requires changes to .rst files it contains to not make
Sphinx barf. For now, I just include the plain tool in the repo.